Refactored tests to separate files
